unified-doc-intelligence-deploy/requirements.txt
# Unified Engineering Document Intelligence Platform - Dependencies
# Install: pip install -r requirements.txt

# PostgreSQL driver
psycopg2-binary>=2.9.9

# Environment variable loading
python-dotenv>=1.0.0

# Progress bars
tqdm>=4.66.0

# Table formatting for reports
tabulate>=0.9.0

# ============================================
# Phase B: PDF/DXF Extraction Dependencies
# ============================================

# PDF extraction (PyMuPDF)
PyMuPDF>=1.23.0

# DXF file parsing
ezdxf>=1.1.0

# Backblaze B2 cloud storage
b2sdk>=1.24.0

# Temporary file handling (stdlib, but good to note)
# tempfile - built-in

# ============================================
# Phase C: Search API Dependencies
# ============================================

# FastAPI web framework
fastapi>=0.109.0

# ASGI server
uvicorn[standard]>=0.27.0

# Form/file handling for FastAPI
python-multipart>=0.0.6

# Pydantic for data validation (comes with FastAPI, but explicit)
pydantic>=2.5.0
